Add merge sort, algorithm modules, and UI/UX improvements
