+++ package.json
@@ -18,7 +18,7 @@
     "react-redux": "^7.2.0",
     "react-scripts": "3.4.1",
     "redux": "^4.0.5",
-    "typescript": "^3.7.5"
+    "typescript": "^4.9.5"
   },
   "scripts": {
     "predeploy": "npm run build",
