+++ src/App.css
@@ -2,6 +2,8 @@
   text-align: center;
   width: 100%;
   height: 100%;
+  background: #f6f8fc;
+  color: #0f172a;
 }
 
 .App-logo {
@@ -51,7 +53,11 @@ html {
 
 body {
   width: 100%;
-  height: 100%
+  height: 100%;
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen,
+    Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
+  background: #f6f8fc;
 }
 
 head {
