+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import logo from './logo.svg';
 import './App.css';
-import VizContainer from './SortingVisualizer/VizContainer'
+import VizContainer from './SortingVisualizer/VizContainer';
 import Header from './Header/header';
 
 const App: React.FC = () => {
@@ -11,6 +11,6 @@ const App: React.FC = () => {
       <VizContainer />
     </div>
   );
-}
+};
 
 export default App;
