+++ src/SortingVisualizer/VizContainer.css
@@ -13,15 +13,19 @@
 .viz-container {
   height: calc(100% - 70px);
   width: 100%;
-  background: white;
+  background: linear-gradient(180deg, #f8fafc 0%, #eef2ff 100%);
   display: flex;
   flex-direction: row;
   align-items: flex-end;
   justify-content: center;
+  padding: 14px 14px 0;
+  box-sizing: border-box;
 }
 
 .viz-bar {
   width: 10px;
   display: flex;
   margin: 0 1px;
+  border-radius: 3px 3px 0 0;
+  transition: height 120ms ease, background-color 120ms ease;
 }
